// src/models/Usuario.ts
export interface Usuario {
    nomeCompleto: string;
    nomeUsuario: string;
    email: string;
    senha: string;
    administrador: string;
    cpf: string;
    telefone: string;
    razaoSocial: string;
    cnpj: string;
    addressLine: string;
    city: string;
    stateRegion: string;
    zipCode: string;
    country: string;
    cardNumber: string;
    expiryDate: string;
}